# src-main/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AttemptStatus": {
        "enum": [
          "submitted",
          "completed"
        ],
        "title": "AttemptStatus",
        "type": "string"
      },
...
      "GenerateTasksRequest": {
        "additionalProperties": false,
        "properties": {
          "allowed_task_types": {
            "items": {
              "$ref": "#/components/schemas/TaskType"
            },
            "title": "Allowed Task Types",
            "type": "array"
          },
          "difficulty_levels": {
            "items": {
              "maxLength": 30,
              "minLength": 1,
              "type": "string"
            },
            "title": "Difficulty Levels",
            "type": "array"
          },
          "learning_outcome_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Learning Outcome Id",
            "type": "string"
          },
          "learning_outcome_text": {
            "minLength": 1,
            "title": "Learning Outcome Text",
            "type": "string"
          },
          "module_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Id"
          },
          "task_count": {
            "default": 3,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Task Count",
            "type": "integer"
          }
        },
        "required": [
          "learning_outcome_id",
          "learning_outcome_text",
          "allowed_task_types",
          "difficulty_levels"
        ],
        "title": "GenerateTasksRequest",
        "type": "object"
      },
      "GeneratedTaskRead": {
        "additionalProperties": false,
        "properties": {
          "difficulty": {
            "minLength": 1,
            "title": "Difficulty",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "instructions": {
            "minLength": 1,
            "title": "Instructions",
            "type": "string"
          },
          "learning_outcome_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Learning Outcome Id",
            "type": "string"
          },
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "source_references": {
            "items": {
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "title": "Source References",
            "type": "array"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "prompt",
          "instructions",
          "task_type",
          "difficulty",
          "learning_outcome_id",
          "source_references"
        ],
        "title": "GeneratedTaskRead",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LatestAttemptSummary": {
        "additionalProperties": false,
        "properties": {
          "attempt_number": {
            "title": "Attempt Number",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "attempt_number",
          "status",
          "score",
          "submitted_at"
        ],
        "title": "LatestAttemptSummary",
        "type": "object"
      },
...
      "TaskChoice": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text"
        ],
        "title": "TaskChoice",
        "type": "object"
      },
      "TaskCreate": {
        "additionalProperties": false,
        "properties": {
          "difficulty": {
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "title": "Difficulty",
            "type": "string"
          },
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "expected_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Answer"
          },
          "instructions": {
            "minLength": 1,
            "title": "Instructions",
            "type": "string"
          },
          "learning_outcome_id": {
            "maxLength": 36,
            "minLength": 36,
            "title": "Learning Outcome Id",
            "type": "string"
          },
          "marking_criteria": {
            "additionalProperties": true,
            "title": "Marking Criteria",
            "type": "object"
          },
          "module_id": {
            "maxLength": 36,
            "minLength": 36,
            "title": "Module Id",
            "type": "string"
          },
          "points": {
            "default": 100,
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Points",
            "type": "integer"
          },
          "position": {
            "exclusiveMinimum": 0.0,
            "title": "Position",
            "type": "integer"
          },
          "prerequisite_task_ids": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisite Task Ids",
            "type": "array"
          },
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "source_references": {
            "items": {
              "type": "string"
            },
            "title": "Source References",
            "type": "array"
          },
          "starter_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starter Code"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "module_id",
          "learning_outcome_id",
          "title",
          "prompt",
          "instructions",
          "task_type",
          "difficulty",
          "position"
        ],
        "title": "TaskCreate",
        "type": "object"
      },
...
      "TaskRead": {
        "additionalProperties": false,
        "properties": {
          "access_status": {
            "enum": [
              "locked",
              "available",
              "in_progress",
              "completed"
            ],
            "title": "Access Status",
            "type": "string"
          },
          "attempt_count": {
            "default": 0,
            "title": "Attempt Count",
            "type": "integer"
          },
          "choices": {
            "items": {
              "$ref": "#/components/schemas/TaskChoice"
            },
            "title": "Choices",
            "type": "array"
          },
          "course_id": {
            "title": "Course Id",
            "type": "string"
          },
          "difficulty": {
            "title": "Difficulty",
            "type": "string"
          },
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "instructions": {
            "title": "Instructions",
            "type": "string"
          },
          "latest_attempt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestAttemptSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Score"
          },
          "learning_outcome_id": {
            "title": "Learning Outcome Id",
            "type": "string"
          },
          "module_id": {
            "title": "Module Id",
            "type": "string"
          },
          "module_title": {
            "title": "Module Title",
            "type": "string"
          },
          "points": {
            "title": "Points",
            "type": "integer"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "prerequisite_task_ids": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisite Task Ids",
            "type": "array"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "source_references": {
            "items": {
              "type": "string"
            },
            "title": "Source References",
            "type": "array"
          },
          "starter_circuit": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starter Circuit"
          },
          "starter_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starter Code"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "prompt",
          "instructions",
          "task_type",
          "difficulty",
          "points",
          "position",
          "starter_code",
          "due_at",
          "course_id",
          "module_id",
          "module_title",
          "learning_outcome_id",
          "source_references",
          "prerequisite_task_ids",
          "access_status"
        ],
        "title": "TaskRead",
        "type": "object"
      },
      "TaskType": {
        "enum": [
          "multiple_choice",
          "multiple_answer",
          "short_answer",
          "code_explanation",
          "code_completion",
          "quantum_circuit",
          "quiz",
          "code",
          "circuit"
        ],
        "title": "TaskType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "QuantumLearn API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}/tasks": {
      "get": {
        "operationId": "list_tasks_api_v1_courses__course_id__tasks_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "quantumlearn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quantumlearn Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "title": "Response List Tasks Api V1 Courses  Course Id  Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "learning management"
        ]
      },
      "post": {
        "operationId": "create_task_api_v1_courses__course_id__tasks_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "quantumlearn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quantumlearn Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task",
        "tags": [
          "learning management"
        ]
      }
    },
    "/api/v1/courses/{course_id}/tasks/generate": {
      "post": {
        "operationId": "generate_tasks_api_v1_courses__course_id__tasks_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "quantumlearn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quantumlearn Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateTasksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeneratedTaskRead"
                  },
                  "title": "Response Generate Tasks Api V1 Courses  Course Id  Tasks Generate Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Tasks",
        "tags": [
          "task generation"
        ]
      }
    },
...
  }
}